+++ libs/entries.js
@@ -256,7 +256,9 @@ module.exports = {
     return fs.statAsync(fpath).then((st) => {
       if (st.isFile()) {
         return self.makePersistent(entryPath, contents).then(() => {
-          return self.updateCache(entryPath)
+          return self.updateCache(entryPath).then(entry => {
+            return search.add(entry)
+          })
         })
       } else {
         return Promise.reject(new Error('Entry does not exist!'))