+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Offline mode (no remote git sync) can now be enabled by setting `git: false` in config.yml
+- Improved search engine (Now using search-index engine instead of MongoDB text search)
+
+### Changed
+- Cache is now flushed when starting / restarting the server
 
 ## [v1.0-beta.4] - 2017-02-11
 ### Fixed